♻️  refactor(examples/election/src/): moved config generation into its own files; fixed small bugs >>> ⏰ 3h
config is loaded in main.py with argument, otherwise gen_config.py is run that writes a config to disk that is then the one loaded in main.py. This cleaves out the config specification from the rest of the codebase. The config now contains most (all?) modifiable settings, e.g. the custom call to action.