♻️  refactor(concordia_utils.py): removed duplicated class defs by adding name field and get_component_name >>> ⏰ 1h
components are stored in a dictionary with class names as keys. We have components that are distinct instances of the same class. We had duplicated the classes to fit that. This commit generalizes the get key function from a get_class_name function to a get_component_name function that checks if the component has a name attribute and uses that instead of the class name.